Category filter feature

+++ YellowNotes.Core/Services/NoteService.cs
@@ -1,12 +1,12 @@
-using System;
-using System.Collections.Generic;
 using System.Linq;
+using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
 using AutoMapper;
 using YellowNotes.Core.Dtos;
 using YellowNotes.Core.Models;
 using YellowNotes.Core.Repositories;
+using YellowNotes.Core.Utility;
 
 namespace YellowNotes.Core.Services
 {
@@ -26,35 +26,93 @@ namespace YellowNotes.Core.Services
             CancellationToken cancellationToken)
         {
             var mappedNote = mapper.Map<Note>(note);
-
             var result = await repository.CreateNote(mappedNote, email, cancellationToken);
             return mapper.Map<NoteDto>(result);
         }
 
-        public async Task<object> GetNote(int noteId, string email,
+        public async Task<ResultHandler> GetNote(int noteId, string email,
             CancellationToken cancellationToken)
         {
             var note = await repository.GetNote(noteId, email, cancellationToken);
-            return mapper.Map<NoteDto>(note);
+
+            if (note == null)
+            {
+                return new ResultHandler(HttpStatusCode.NotFound);
+            }
+            else if (note.User.Email != email)
+            {
+                return new ResultHandler(HttpStatusCode.Unauthorized,
+                    "Requested resource is not available!");
+            }
+            return new ResultHandler(HttpStatusCode.OK, mapper.Map<NoteDto>(note));
         }
 
-        public async Task<Tuple<int, IEnumerable<NoteDto>>> GetNotes(int takeCount, int skipCount,
-            string email, CancellationToken cancellationToken)
-        {
-            var notes = await repository.GetNotes(takeCount, skipCount, email, cancellationToken);
-            return Tuple.Create(notes.Item1, notes.Item2.Select(x => mapper.Map<NoteDto>(x)));
-        }
-
-        public async Task<object> UpdateNote(NoteDto note, string email, 
+        public async Task<NotesDto> GetNotes(NoteQueryDto query, string email,
             CancellationToken cancellationToken)
         {
-            return await repository.UpdateNote(note, email, cancellationToken);
+            var notesData = await repository.GetNotes(query, email, cancellationToken);
+            notesData.Notes = notesData.Notes
+                .Where(x => IsCategorySelected(x.CategoryId, query.CategoryIds))
+                .ToList();
+
+            return new NotesDto
+            {
+                Count = notesData.Count,
+                Notes = notesData.Notes.Select(x => mapper.Map<NoteDto>(x))
+            };
         }
 
-        public async Task<object> DeleteNote(int noteId, string email,
+        public async Task<ResultHandler> UpdateNote(NoteDto note, string email,
             CancellationToken cancellationToken)
         {
-            return await repository.DeleteNote(noteId, email, cancellationToken);
+            var record = await repository.GetNote(note.NoteId, email, cancellationToken);
+
+            if (record == null)
+            {
+                return new ResultHandler(HttpStatusCode.NotFound);
+            }
+            else if (record.User.Email != email)
+            {
+                return new ResultHandler(HttpStatusCode.Unauthorized,
+                    "Requested resource is not available!");
+            }
+
+            var success = await repository.UpdateNote(note, email, cancellationToken);
+
+            return success
+                ? new ResultHandler(HttpStatusCode.NoContent)
+                : new ResultHandler(HttpStatusCode.UnprocessableEntity, "Failed to update note");
+        }
+
+        public async Task<ResultHandler> DeleteNote(int noteId, string email,
+            CancellationToken cancellationToken)
+        {
+            var note = await repository.GetNote(noteId, email, cancellationToken);
+
+            if (note == null)
+            {
+                return new ResultHandler(HttpStatusCode.NotFound);
+            }
+            else if (note.User.Email != email)
+            {
+                return new ResultHandler(HttpStatusCode.Unauthorized,
+                    "Requested resource is not available!");
+            }
+
+            var success = await repository.DeleteNote(noteId, email, cancellationToken);
+
+            return success
+                ? new ResultHandler(HttpStatusCode.NoContent)
+                : new ResultHandler(HttpStatusCode.NotFound);
+        }
+
+        private bool IsCategorySelected(int? categoryId, int[] categoryIds)
+        {
+            if (categoryIds.Length > 0)
+            {
+                return categoryId.HasValue ? categoryIds.Contains(categoryId.Value) : false;
+            }
+            return true;
         }
     }
 }
